refactor(db): extract remote_connection into db/remote.rs
